pyproject.toml
[project]
name = "llm-web-navigation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "huggingface-hub>=1.14.0",
    "pillow>=12.2.0",
    "playwright>=1.59.0",
    "playwright-stealth>=2.0.3",
    "ultralytics>=8.4.50",
    "pi-heif==1.3.0",
    "torch>=2.12.0",
    "transformers>=5.8.1",
    "qwen-vl-utils>=0.0.14",
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
